# pkg/client/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: info@green-ecolution.de
    name: Green Ecolution
    url: https://green-ecolution.de
  description: This is the API for the Green Ecolution Management System.
  license:
    name: AGPL
    url: https://raw.githubusercontent.com/green-ecolution/green-ecolution-management/develop/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Green Space Management API
  version: develop
servers:
- url: /
paths:
  /v1/cluster:
    get:
      description: Get all tree clusters
      operationId: get-all-tree-clusters
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: string
      - description: Provider
        in: query
        name: provider
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeClusterList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get all tree clusters
      tags:
      - Tree Cluster
    post:
      description: Create tree cluster
      operationId: create-tree-cluster
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/TreeClusterCreate'
        description: Tree Cluster Create Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeCluster'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Create tree cluster
      tags:
      - Tree Cluster
      x-codegen-request-body-name: body
  /v1/cluster/{cluster_id}:
    delete:
      description: Delete tree cluster
      operationId: delete-tree-cluster
      parameters:
      - description: Tree Cluster ID
        in: path
        name: cluster_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Delete tree cluster
      tags:
      - Tree Cluster
    get:
      description: Get tree cluster by ID
      operationId: get-tree-cluster-by-id
      parameters:
      - description: Tree Cluster ID
        in: path
        name: cluster_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeCluster'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get tree cluster by ID
      tags:
      - Tree Cluster
    put:
      description: Update tree cluster
      operationId: update-tree-cluster
      parameters:
      - description: Tree Cluster ID
        in: path
        name: cluster_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/TreeClusterUpdate'
        description: Tree Cluster Update Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeCluster'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Update tree cluster
      tags:
      - Tree Cluster
      x-codegen-request-body-name: body
...
  /v1/tree:
    get:
      description: Get all trees
      operationId: get-all-trees
      parameters:
      - description: Page
        in: query
        name: page
        schema:
          type: string
      - description: Limit
        in: query
        name: limit
        schema:
          type: string
      - description: Provider
        in: query
        name: provider
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeList'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get all trees
      tags:
      - Tree
    post:
      description: Create tree
      operationId: create-tree
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/TreeCreate'
        description: Tree to create
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tree'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Create tree
      tags:
      - Tree
      x-codegen-request-body-name: body
  /v1/tree/{tree_id}:
    delete:
      description: Delete tree
      operationId: delete-tree
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        schema:
          type: integer
      responses:
        "204":
          content: {}
          description: No Content
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Delete tree
      tags:
      - Tree
    get:
      description: Get tree by ID
      operationId: get-trees
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tree'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Get tree by ID
      tags:
      - Tree
    put:
      description: Update tree
      operationId: update-tree
      parameters:
      - description: Tree ID
        in: path
        name: tree_id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          '*/*':
            schema:
              $ref: '#/components/schemas/TreeUpdate'
        description: Tree to update
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tree'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      security:
      - Keycloak: []
      summary: Update tree
      tags:
      - Tree
      x-codegen-request-body-name: body
...
components:
  schemas:
...
    HTTPError:
      example:
        path: path
        code: 7
        method: method
        error: error
      properties:
        code:
          type: integer
        error:
          type: string
        method:
          type: string
        path:
          type: string
      required:
      - code
      - error
      - method
      - path
      type: object
...
    SoilCondition:
      enum:
      - schluffig
      - sandig
      - lehmig
      - tonig
      - unknown
      type: string
      x-enum-varnames:
      - TreeSoilConditionSchluffig
      - TreeSoilConditionSandig
      - TreeSoilConditionLehmig
      - TreeSoilConditionTonig
      - TreeSoilConditionUnknown
    Tree:
      example:
        latitude: 2.3021358869347655
        created_at: created_at
        description: description
        tree_cluster_id: 7
        additional_information:
          key: ""
        number: number
        readonly: true
        updated_at: updated_at
        watering_status: good
        provider: provider
        species: species
        planting_year: 9
        sensor:
          additional_information:
            key: ""
          latest_data:
            updated_at: updated_at
            temperature: 4.145608029883936
            created_at: created_at
            humidity: 2.027123023002322
            watermarks:
            - centibar: 7
              depth: 1
              resistance: 1
            - centibar: 7
              depth: 1
              resistance: 1
            battery: 3.616076749251911
          updated_at: updated_at
          provider: provider
          latitude: 1.4894159098541704
          created_at: created_at
          id: id
          longitude: 6.84685269835264
          status: online
        id: 5
        longitude: 7.061401241503109
      properties:
        additional_information:
          additionalProperties: true
          type: object
        created_at:
          type: string
        description:
          type: string
        id:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        number:
          type: string
        planting_year:
          type: integer
        provider:
          type: string
        readonly:
          description: "Images              []*ImageResponse `json:\"images\"`"
          type: boolean
        sensor:
          $ref: '#/components/schemas/Sensor'
        species:
          type: string
        tree_cluster_id:
          type: integer
        updated_at:
          type: string
        watering_status:
          $ref: '#/components/schemas/WateringStatus'
      required:
      - additional_information
      - created_at
      - description
      - id
      - latitude
      - longitude
      - number
      - planting_year
      - provider
      - readonly
      - species
      - updated_at
      - watering_status
      type: object
    TreeCluster:
      example:
        address: address
        latitude: 6.027456183070403
        created_at: created_at
        description: description
        last_watered: last_watered
        trees:
        - latitude: 2.3021358869347655
          created_at: created_at
          description: description
          tree_cluster_id: 7
          additional_information:
            key: ""
          number: number
          readonly: true
          updated_at: updated_at
          watering_status: good
          provider: provider
          species: species
          planting_year: 9
          sensor:
            additional_information:
              key: ""
            latest_data:
              updated_at: updated_at
              temperature: 4.145608029883936
              created_at: created_at
              humidity: 2.027123023002322
              watermarks:
              - centibar: 7
                depth: 1
                resistance: 1
              - centibar: 7
                depth: 1
                resistance: 1
              battery: 3.616076749251911
            updated_at: updated_at
            provider: provider
            latitude: 1.4894159098541704
            created_at: created_at
            id: id
            longitude: 6.84685269835264
            status: online
          id: 5
          longitude: 7.061401241503109
        - latitude: 2.3021358869347655
          created_at: created_at
          description: description
          tree_cluster_id: 7
          additional_information:
            key: ""
          number: number
          readonly: true
          updated_at: updated_at
          watering_status: good
          provider: provider
          species: species
          planting_year: 9
          sensor:
            additional_information:
              key: ""
            latest_data:
              updated_at: updated_at
              temperature: 4.145608029883936
              created_at: created_at
              humidity: 2.027123023002322
              watermarks:
              - centibar: 7
                depth: 1
                resistance: 1
              - centibar: 7
                depth: 1
                resistance: 1
              battery: 3.616076749251911
            updated_at: updated_at
            provider: provider
            latitude: 1.4894159098541704
            created_at: created_at
            id: id
            longitude: 6.84685269835264
            status: online
          id: 5
          longitude: 7.061401241503109
        soil_condition: schluffig
        additional_information:
          key: ""
        archived: true
        updated_at: updated_at
        watering_status: null
        provider: provider
        name: name
        moisture_level: 5.962133916683182
        id: 0
        region:
          name: name
          id: 5
        longitude: 1.4658129805029452
      properties:
        additional_information:
          additionalProperties: true
          type: object
        address:
          type: string
        archived:
          type: boolean
        created_at:
          type: string
        description:
          type: string
        id:
          type: integer
        last_watered:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        moisture_level:
          type: number
        name:
          type: string
        provider:
          type: string
        region:
          $ref: '#/components/schemas/Region'
        soil_condition:
          $ref: '#/components/schemas/SoilCondition'
        trees:
          items:
            $ref: '#/components/schemas/Tree'
          type: array
        updated_at:
          type: string
        watering_status:
          $ref: '#/components/schemas/WateringStatus'
      required:
      - additional_information
      - address
      - archived
      - created_at
      - description
      - id
      - latitude
      - longitude
      - moisture_level
      - name
      - provider
      - soil_condition
      - updated_at
      - watering_status
      type: object
    TreeClusterCreate:
      example:
        additional_information:
          key: ""
        address: address
        tree_ids:
        - 0
        - 0
        provider: provider
        name: name
        description: description
        soil_condition: schluffig
      properties:
        additional_information:
          additionalProperties: true
          type: object
        address:
          type: string
        description:
          type: string
        name:
          type: string
        provider:
          type: string
        soil_condition:
          $ref: '#/components/schemas/SoilCondition'
        tree_ids:
          items:
            type: integer
          type: array
      required:
      - address
      - description
      - name
      - soil_condition
      - tree_ids
      type: object
...
    TreeClusterList:
      example:
        pagination:
          next_page: 9
          total_records: 4
          total_pages: 2
          prev_page: 3
          current_page: 7
        data:
        - address: address
          latitude: 6.027456183070403
          created_at: created_at
          description: description
          last_watered: last_watered
          soil_condition: schluffig
          archived: true
          tree_ids:
          - 2
          - 2
          updated_at: updated_at
          watering_status: good
          name: name
          moisture_level: 5.962133916683182
          id: 0
          region:
            name: name
            id: 5
          longitude: 1.4658129805029452
        - address: address
          latitude: 6.027456183070403
          created_at: created_at
          description: description
          last_watered: last_watered
          soil_condition: schluffig
          archived: true
          tree_ids:
          - 2
          - 2
          updated_at: updated_at
          watering_status: good
          name: name
          moisture_level: 5.962133916683182
          id: 0
          region:
            name: name
            id: 5
          longitude: 1.4658129805029452
      properties:
        data:
          items:
            $ref: '#/components/schemas/TreeClusterInList'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      type: object
    TreeClusterUpdate:
      example:
        additional_information:
          key: ""
        address: address
        tree_ids:
        - 0
        - 0
        provider: provider
        name: name
        description: description
        soil_condition: schluffig
      properties:
        additional_information:
          additionalProperties: true
          type: object
        address:
          type: string
        description:
          type: string
        name:
          type: string
        provider:
          type: string
        soil_condition:
          $ref: '#/components/schemas/SoilCondition'
        tree_ids:
          items:
            type: integer
          type: array
      required:
      - address
      - description
      - name
      - soil_condition
      - tree_ids
      type: object
    TreeCreate:
      example:
        sensor_id: sensor_id
        additional_information:
          key: ""
        number: number
        readonly: true
        provider: provider
        species: species
        latitude: 0.8008281904610115
        description: description
        planting_year: 1
        tree_cluster_id: 5
        longitude: 6.027456183070403
      properties:
        additional_information:
          additionalProperties: true
          type: object
        description:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        number:
          type: string
        planting_year:
          type: integer
        provider:
          type: string
        readonly:
          type: boolean
        sensor_id:
          type: string
        species:
          type: string
        tree_cluster_id:
          type: integer
      required:
      - description
      - latitude
      - longitude
      - number
      - planting_year
      - readonly
      - species
      type: object
    TreeList:
      example:
        pagination:
          next_page: 9
          total_records: 4
          total_pages: 2
          prev_page: 3
          current_page: 7
        data:
        - latitude: 2.3021358869347655
          created_at: created_at
          description: description
          tree_cluster_id: 7
          additional_information:
            key: ""
          number: number
          readonly: true
          updated_at: updated_at
          watering_status: good
          provider: provider
          species: species
          planting_year: 9
          sensor:
            additional_information:
              key: ""
            latest_data:
              updated_at: updated_at
              temperature: 4.145608029883936
              created_at: created_at
              humidity: 2.027123023002322
              watermarks:
              - centibar: 7
                depth: 1
                resistance: 1
              - centibar: 7
                depth: 1
                resistance: 1
              battery: 3.616076749251911
            updated_at: updated_at
            provider: provider
            latitude: 1.4894159098541704
            created_at: created_at
            id: id
            longitude: 6.84685269835264
            status: online
          id: 5
          longitude: 7.061401241503109
        - latitude: 2.3021358869347655
          created_at: created_at
          description: description
          tree_cluster_id: 7
          additional_information:
            key: ""
          number: number
          readonly: true
          updated_at: updated_at
          watering_status: good
          provider: provider
          species: species
          planting_year: 9
          sensor:
            additional_information:
              key: ""
            latest_data:
              updated_at: updated_at
              temperature: 4.145608029883936
              created_at: created_at
              humidity: 2.027123023002322
              watermarks:
              - centibar: 7
                depth: 1
                resistance: 1
              - centibar: 7
                depth: 1
                resistance: 1
              battery: 3.616076749251911
            updated_at: updated_at
            provider: provider
            latitude: 1.4894159098541704
            created_at: created_at
            id: id
            longitude: 6.84685269835264
            status: online
          id: 5
          longitude: 7.061401241503109
      properties:
        data:
          items:
            $ref: '#/components/schemas/Tree'
          type: array
        pagination:
          $ref: '#/components/schemas/Pagination'
      required:
      - data
      type: object
...
  securitySchemes:
    Keycloak:
      flows:
        authorizationCode:
          authorizationUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/auth
          scopes: {}
          tokenUrl: https://auth.green-ecolution.de/realms/green-ecolution-dev/protocol/openid-connect/token
      type: oauth2
...